Consolidate projects/adjust module
